Add student results per subject and class

+++ FimiAppLibrary/Models/StudentResultsModel.cs
@@ -10,16 +10,16 @@
         public string FirstName { get; set; }
         public string MiddleName { get; set; }
         public string Surname { get; set; }
-        public int English { get; set; }
-        public int Kiswahili { get; set; }
-        public int Mathematics { get; set; }
-        public int Physics { get; set; }
-        public int Chemistry { get; set; }
-        public int Biology { get; set; }
-        public int HistoryGovernment { get; set; }
-        public int Cre { get; set; }
-        public int HomeScience { get; set; }
-        public int Agriculture { get; set; }
-        public int BusinessStudies { get; set; }
+        public double English { get; set; }
+        public double Kiswahili { get; set; }
+        public double Mathematics { get; set; }
+        public double Physics { get; set; }
+        public double Chemistry { get; set; }
+        public double Biology { get; set; }
+        public double HistoryGovernment { get; set; }
+        public double Cre { get; set; }
+        public double HomeScience { get; set; }
+        public double Agriculture { get; set; }
+        public double BusinessStudies { get; set; }
     }
 }
